REPO-5613 Extract common auth config to external classes
