+++ alfresco-java-rest-api/alfresco-java-rest-api-common/src/main/java/org/alfresco/rest/sdk/feign/config/BasicAuthConfiguration.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2021-2021 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.alfresco.rest.sdk.feign.config;
+
+import feign.auth.BasicAuthRequestInterceptor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration for basic authentication. It configures a {@link BasicAuthRequestInterceptor} that uses the credentials configured in the properties
+ * <code>content.service.security.basicAuth.username</code> and <code>content.service.security.basicAuth.password</code>.
+ */
+@Configuration
+@ConditionalOnProperty("content.service.security.basicAuth.username")
+public class BasicAuthConfiguration {
+
+    @Value("${content.service.security.basicAuth.username:#{null}}")
+    private String basicAuthUsername;
+    @Value("${content.service.security.basicAuth.password:#{null}}")
+    private String basicAuthPassword;
+
+    @Bean
+    @ConditionalOnProperty("content.service.security.basicAuth.username")
+    public BasicAuthRequestInterceptor basicAuthRequestInterceptor() {
+        return new BasicAuthRequestInterceptor(this.basicAuthUsername, this.basicAuthPassword);
+    }
+}
